+++ app/src/main/AndroidManifest.xml
@@ -16,13 +16,13 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".ui.order.OrderHistoryActivity"
+            android:name=".ui.order.history.OrderHistoryActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.order.OrderDetailActivity"
+            android:name=".ui.order.detail.OrderDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.order.OrderActivity"
+            android:name=".ui.order.main.OrderActivity"
             android:exported="false" />
         <activity
             android:name=".ui.serversetting.ServerSettingActivity"
