--- app/src/main/java/woowacourse/shopping/ui/order/OrderDetailActivity.kt
@@ -1,12 +1,0 @@
-package woowacourse.shopping.ui.order
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import woowacourse.shopping.R
-
-class OrderDetailActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_order_detail)
-    }
-}
