+++ app/src/main/res/layout/activity_order_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.order.OrderDetailActivity">
+    tools:context=".ui.order.detail.OrderDetailActivity">
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
