+++ app/src/main/res/layout/activity_order_history.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.order.OrderHistoryActivity">
+    tools:context=".ui.order.history.OrderHistoryActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
